Split node-migrate & the filestore Reusable tests change migration script
